Elaboração dos serviços de Address e Agent.

+++ Assembly.Projecto.Final.Services/Services/AddressService.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -23,13 +24,14 @@ namespace Assembly.Projecto.Final.Services.Services
 
         public AddressDto Add(CreateAddressDto createAddressDto)
         {
-            var address = _mapper.Map<Address>(createAddressDto);
-
             Address addedAddress;
 
             using (_unitOfWork) 
             {
                 _unitOfWork.BeginTransaction();
+
+                var address = Address.Create(createAddressDto.Street, createAddressDto.City, createAddressDto.Country,
+                    createAddressDto.PostalCode);
 
                 addedAddress = _unitOfWork.AddressRepository.Add(address);
 
@@ -41,22 +43,20 @@ namespace Assembly.Projecto.Final.Services.Services
 
         public AddressDto Delete(AddressDto addressDto)
         {
-            var address = _mapper.Map<Address>(addressDto);
-
             Address deletedAddress;
 
             using (_unitOfWork) 
             {
                 _unitOfWork.BeginTransaction();
 
-                var foundedAddress = _unitOfWork.AddressRepository.GetById(address.Id);
+                var foundedAddress = _unitOfWork.AddressRepository.GetById(addressDto.Id);
 
                 if (foundedAddress == null) 
                 {
-                    throw new ArgumentNullException(nameof(foundedAddress), "Not found");
+                    throw new ArgumentNullException(nameof(foundedAddress), "Não foi encontrado.");
                 }
 
-                deletedAddress =_unitOfWork.AddressRepository.Delete(address);
+                deletedAddress =_unitOfWork.AddressRepository.Delete(foundedAddress);
 
                 _unitOfWork.Commit();
             }
@@ -67,22 +67,72 @@ namespace Assembly.Projecto.Final.Services.Services
 
         public AddressDto Delete(int id)
         {
-            return _unitOfWork.AddressRepository.Delete(id);
+            Address deletedAddress;
+
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAddress = _unitOfWork.AddressRepository.GetById(id);
+
+                if(foundedAddress == null) 
+                {
+                    throw new ArgumentNullException(nameof(foundedAddress), "Não foi encontrado.");
+                }
+
+                deletedAddress =  _unitOfWork.AddressRepository.Delete(id)
+
+                _unitOfWork.Commit();
+
+            }
+
+            return _mapper.Map<AddressDto>(deletedAddress);
         }
 
         public List<AddressDto> GetAll()
         {
-            return _unitOfWork.AddressRepository.GetAll();
+            var list = new List<AddressDto>();
+
+            foreach(var address in _unitOfWork.AddressRepository.GetAll()) 
+            {
+                var addressDto = _mapper.Map<AddressDto>(address);
+
+                list.Add(addressDto);
+            }
+
+            return list;
         }
 
         public AddressDto GetById(int id)
         {
-            return _unitOfWork.AddressRepository.GetById(id);
+            var address =_unitOfWork.AddressRepository.GetById(id);
+
+            return _mapper.Map<AddressDto>(address);
         }
 
         public AddressDto Update(AddressDto addressDto)
         {
-             return _unitOfWork.AddressRepository.Update(address);
+            Address updatedAddress;
+            
+            using (_unitOfWork) 
+            {
+                _unitOfWork.BeginTransaction();
+
+                var foundedAddress = _unitOfWork.AddressRepository.GetById(addressDto.Id);  
+                
+                if(foundedAddress == null) 
+                {
+                    throw new ArgumentNullException(nameof(foundedAddress), "Não foi encontrado.");
+                }
+
+                foundedAddress.Update(addressDto.Street, addressDto.City, addressDto.Country, addressDto.PostalCode);
+
+                updatedAddress = _unitOfWork.AddressRepository.Update(foundedAddress);
+
+                _unitOfWork.Commit();
+            }
+
+            return _mapper.Map<AddressDto>(updatedAddress);
         }
     }
 }
